backend/pyproject.toml
[project]
name = "renthouse-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.15.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "shapely>=2.0.0",
    "beautifulsoup4>=4.14.3",
    "apscheduler>=3.11.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
